fix: silly mistakes
