+++ src/Settings/DefaultSettings.ts
@@ -1,28 +1,17 @@
-import {type ISettings} from 'Interfaces/Settings';
-import {Feature} from 'Settings/Feature';
 
-// Default data block that is stored in the data file.
-// export const DefaultSettings: ISettings = {
-//   features: Feature.settingsFlags,
-//   generalSettings: {
-//     refreshDebounce: 2500,
-//     settingsVersion: 1,
-//     internalLoggingConsoleLogLimit: 10,
-//     postRenderFormat: 'micromark',
-//   },
-//   headingOpened: {}, // ;  { 'Documentation and Support': true },
-//   loggingOptions: {
-//     minLevels: {
-//       '': 'info',
-//       Qatt: 'info',
-//     },
-//   },
-// };
-
-export interface IPluginSettings {
+export interface ISqlSettings {
   onStartSqlQueries: string;
 }
 
-export const PluginSettingsDefaults: IPluginSettings = {
+export const SqlSettingsDefaults: ISqlSettings = {
   onStartSqlQueries: 'SELECT * FROM sqlite_master;',
 };
+
+export interface IGeneralSettings {
+  onStartSqlQueries: string;
+}
+
+export const GeneralSettingsDefaults: IGeneralSettings = {
+  onStartSqlQueries: 'SELECT * FROM sqlite_master;',
+};
+
